refacotr: Separando grupos de testes em camadas
