+++ tests/Unit/Auth/AuthTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Auth;
+namespace TestsUnit\Auth;
 
 use App\Models\User;
 use App\Repositories\IUserRepository;
